bot/requirements.txt
discord.py
aiohttp
flask
cryptography
python-dotenv